+++ package.json
@@ -12,6 +12,8 @@
   "author": "Billy Sutomo <billysutomo.53@gmail.com>",
   "dependencies": {
     "moment": "^2.21.0",
+    "react": "^16.4.0",
+    "react-dom": "^16.2.0",
     "react-tooltip": "^3.5.0"
   },
   "devDependencies": {
@@ -27,6 +29,9 @@
     "copy-webpack-plugin": "^4.3.1",
     "css-loader": "0.28.7",
     "dotenv": "4.0.0",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme-to-json": "^3.3.4",
     "eslint": "4.10.0",
     "eslint-config-react-app": "^2.0.1",
     "eslint-loader": "1.9.0",
@@ -45,9 +50,8 @@
     "postcss-loader": "2.0.8",
     "promise": "8.0.1",
     "raf": "3.4.0",
-    "react": "^16.2.0",
     "react-dev-utils": "^4.2.1",
-    "react-dom": "^16.2.0",
+    "react-test-renderer": "^16.4.0",
     "sass-loader": "^6.0.7",
     "style-loader": "0.19.0",
     "sw-precache-webpack-plugin": "0.11.4",
