+++ src/lib/components/Filter/styles.scss
@@ -58,7 +58,7 @@
         text-align: right;
 
         &::placeholder {
-          text-align: right;
+          text-align: center;
         }
       }
     }
